Upgrade to Ruby 3.4.1 and lock rdoc version
This upgrades our Ruby version to the latest, 3.4.1. We also lock
the rdoc version to 6.6.3.1 since that is Ruby's default version,
and it's causing conflicts with the latest rdoc version 6.10.0.
We don't need any features from the latest rdoc version, so for now
we'll just lock it to a known stable version with Ruby.