Add Plaid Items table
This introduces a plaid_items table and model. This will be
responsible for storing Plaid Items, which contain the access_key
to communicate with Plaid's servers.

A Plaid Item object can contain multiple accounts, but will only
belong to one institution. Therefore, we store the institution ID
along with the item.

We add an index for transactions_synced_at and accounts_synced_at
becuase we will have background jobs that run and query for items
that haven't had their transactions or accounts updated in over 8
hours.