Add required plaid columns to account model
This adds columns to the account model to support the new plaid
integration. An Plaid account will always belong to a Plaid Item.
We also rename bank_name to institution_name which more accurately
represents this field's value.

All references to account.bank_name have been updated to
institution_name, except for the transaction serializer, which
currently uses the attribute name "bank". This will be updated in a
future commit since this commit is already very large with the
refactors.