Add Plaid Services API class
This adds a new service class for the Plaid API. This class is
responsible for interacting with the Plaid external service and
error handling.

This currently supports linkTokenCreate, itemPublicTokenExchange,
itemGet, itemRemove, accountsGet, and transactionsSync.

We will most likely need to add investmentsTransactionsGet when we
start to support brokerage accounts. For now, the current library
supports all despository, CD, and credit card accounts.

The transactions_sync method will return the added, modified, and
removed transactions along with the next-cursor. It is the
responsibilty of the caller of this method to update the
PlaidItem.sync_cursor attribute. The API is only responsible for
fetching the data from Plaid, not updating any persistent storage.