Add Plaid ItemCreateService
This adds a new Plaid ItemCreateService which is responsible for
taking the public token, exchanging it for an access token, and
creating a new PlaidItem in the database with the access token and
item_id.

If the item id exists, a record not unique error will be raised. We
attempt to destroy the item on Plaids servers if this happens.

There is a todo item here in that we need to create the background
job that gets the Plaid Item information such as the institution
id. When this job is implemented we will circle back to this
service and add the enqueue logic.