Install and implement sidekiq as queue adapter
This commit installs the sidekiq gem and configures the application
to use it as the active_job queue adapter. We also enable the
cookie store to use the Sidekiq UI.

This also removes the ActiveJob::Status gem as it is not compatible
with Sidekiq jobs.

The CategorizeTransactionsJob and ParseStatementsJob are updated to
use Sidekiq as the queue adapter and removes ActiveJob inheritence.

The logic for ActiveJob::Status was also removed from
CategorizeTransactionsJob. The status endpoint was removed from the
CategorizationJobsController until we can figure out how to return
status from Sidekiq queued jobs.

There are also a few rubocop auto fixes.