Add Plaid GetItemDetailsJob
This implements a background job to fetch item details such as
institution_id from Plaid's api, and saves it to the PlaitItem
record.

We also update the Plaid ItemCreate service to call this new job
after it creates a new PlaidItem record. The institution_id is not
returned in the public token exchange, so we need to fetch it in a
separate API call. The background job allows us to queue this
action and quickly return the API response for exchanging the
token instead of waiting on Plaid for two API calls.