Add Plaid Controller
This creates a new plaid controller which will be utilized by the
Link component in the UI. The Link component is a drop in Plaid
made component that handles the auth flow for users connecting to
their bank. It requires two API calls:
1. create_link_token
2. set_access_token

set_access_token utilizes the ItemCreate service to exchange the
public token for an access token and then creates a new plaid item
record storing the access token.

As of now since this webapp is in development and will not be open
to the public, we are accepting the userId param from the UI. In
the future when we have an authentication mechanism, we will pull
the userId from the user's session and assign the plaid item to the
logged in user.