Add PlaidItem and Account fixtures and associations
This adds the model associations between PlaidItem and Account. The
association is optional as not all accounts will belong to a
PlaidItem.

This also fills in the fixtures we will need for PlaidItems and
Accounts in upcoming tests. We update existing tests that use the
PlaidItem fixtures to reference these new fixtures.

We also update two account data entity tests that were
unfortunately hard coding the number of accounts in the fixtures.
This was a terrible idea and we now fixed it so that the test is
not dependent on the number of accounts.