PlaidItem: Remove plaid item on delete
This adds logic to the PlaidItem model to call the itemRemove API
on PlaidItem destroy. This will stop Plaid from billing us for
unused items.

If the API returns an ITEM_NOT_FOUND error, that means the item was
already removed from Plaid's service and we should proceed with the
local deletion. Any other type of API error will raise a
Plaid::ApiError exception which needs to be handled by all callers
to PlaidItem.destroy.

Tests were added for all edge cases in this scenario.