Add Plaid Sync Account service
This adds a new service responsible for syncing Plaid accounts from
AccountsGet API request.

If the account doesn't exist yet, it will create a new account
record and link it to the relevant Plaid Item. If the account
exists, we will only update two fields if they have changed:
balance and name. All others will remain constant.

It's important to note that if someone links their bank twice,
Plaid sends over a new account id, so we will not know if the
physical account exists already, but the user should know not to
link their accounts twice. We will implement a feature in the
future to merge accounts which combines all the transactions into
one.

Full spec test suite was written to cover all edge cases for this
service.