Refactor Plaid ItemCreate service
This refactors the item create service to clean up the call method.
We move the RecordNotUnique rescue into the create method where it
would be raised. We also stop converting the API response to a hash
and just use the Plaid Response class attributes directly for a
cleaner experience.

This also adds logic to enqueue a fetch_accounts_job after the item
is created. Tests were added to ensure this job is enqueued with
the correct arguments.