Plaid get_item_details_job: institution_name
This adds logic to the Plaid get item details background job to be
able to fetch the institution name. The institution name must be
fetched from a separate API endpoint than the get Item endpoint. So
get_item will first fetch the institution_id, then it will fetch
the institution name using this ID.

If the Item has any accounts, the job will attempt to update the
institution_name on all of the accounts. We do this because the
fetch_accounts job run before or after the get_item_details job, so
it's possible the accounts haven't been synced yet. The
fetch_accounts job will be updated to check if the institution_name
is set on the item and if so, sync it to the accounts upon
creation.