Refactor sync_accounts service and background job
This refactors the sync_accounts background job to move the API
fetching out of the job and into the service. Now, the
sync_accounts service handles all of the logic to sync accounts to
an item which includes data fetching and local updating.

The background job is now only resonsible for identifying the items
that need to be synced and calling the sync_accounts service on
each one.

New logic was added to mark the item's accounts_synced_at column to
now to indicate the accounts were succesfully synced.