Plaid transactions create service
This adds a new service responsible for creating transactions into
an account by reading a Plaid::Transaction object.

Callers to this service should handle an ActiveRecord::RecordNotUnique
error which will occur if this service attempts to create a
transaction with an existing plaid_transaction_id.

The plaid_transaction.authorization_date is favored over the
transaction_date as this is when the actual charge occurred.

A future improvement to this service would be to handle Plaid's
auto categorization which would need to tie into our webapp's
categories & subcategories. Perhaps a future service maps these
values.