Plaid transactions sync service
This adds a new service that is responsible for fetching
transactions from Plaid's API and syncing them to each account. The
Plaid transaction_sync API returns added, modified, and removed
transactions. I'm not quite sure how modified and removed works
yet because sandbox doesn't offer example values. For this reason,
only added is implemented and tested.

This service also marks the PlaidItem's transactions_synced_at and
transaction_sync_cursor attributes when syncing is completed.

We also handle an ActiveRecord::RecordNotUnique error from the
CreateService by logging and skipping the transaction. This
behavior is favored since we want to keep the services idempotent.
If the service is executed twice with the same data, then we won't
create duplicate transactions in the local database.