Plaid Transaction RemoveService
This adds a new service responsible for removing Plaid
transactions. The service takes in an account and array of
Plaid::RemovedTransaction objects. It returns an array of
plaid_transaction_ids that were actually destroyed. If any of the
transaction_ids don't exist, the service will just ignore them and
won't return them in the response.

This also updates the Transaction model to dependent: :destroy any
transaction splits when a parent transaction is destroyed. We need
this logic in place if a plaid transaction is removed and the user
has already split it, then the splits shouldn't be left in tact.
They need to be destroyed as well.