Plaid Transaction SyncService: removed transactions
This commit implements the removed transactions portion of the
Plaid Transactions Sync Service. If any transactions are present in
the removed response of the TransactionSync API call, the
SyncService will call the RemoveService with the account and
transaction IDs.

The service also logs the requested deleted transactions as well as
the ones actually deleted for any necessary future debugging
purposes.