Refactor Transaction serializer for simple fields
This commit refactors the Transaction serializer to simplify the
data returned for the transaction resource. The user object and
account object are removed, and only the userId and accountId are
returned. If the user or account attributes are needed, they must
be joined on the user or account resource from their APIs.