Fix bug in accounts update endpoint
This fixes a bug where the account update endpoint was returning a
null value on success. This is because the endpoint was using a
local variable for the account, but attempting to return an
instance variable of the same name.

We add tests to make sure the endpoint response is correct which
would catch this issue in the future.