Plaid AccountTypeMapper model
This commit adds a new model responsible for mapping Plaid account
types to our app's account types. We don't need the granular
account subtypes that Plaid provides. We only need to store whether
the account is an asset or liability, and if it's cash,
investments, credit cards, or loans.

There is an other type that Plaid provides, and im unsure which
accounts fall under that category. For now, im adding it to Assets
but if in the future we determine it needs to be updated, we can
certainly do that.