Plaid Sync Accounts: use Account Type Mapper
This updates the Plaid::SyncAccounts service to use the
AccountTypeMapper model for the account type and subtype values.

The AccountTypeMapper will raise an InvalidAccountType error if the
Plaid account type is unknown. So we'll log when this happens but
continue to create the remaining accounts instead of stopping
account creation.

We also expand the tests to ensure we are testing the new
mappings.