Plaid SyncAccounts: sync institution_name from item
This commit updates the Plaid::SyncAccounts service to sync the
institution_name from the associated Plaid Item if it's set.

The Get Item Details Job is triggered at the same time as the
SyncAccounts job. If SyncAccounts gets to the creation step before
Get Item Details Job, then it will create the account with out an
institution_name. In this case, the Get Item Details Job will sync
the institution_name to the Account once it fetches the data.

If the Get Item Details Job fetches the institution_name before the
Sync Accounts service creates the accounts, then the SyncAccounts
job will pull the institution_name from the associated PlaidItem.

The reason we can't only have an institution_id on the Plaid Item
is that some accounts won't be linked through Plaid and therefore
won't have a plaid_item association.