Add Sidekiq::Status gem and configuration
This adds the Sidekiq Status gem to our application, and configures
the necessary initialization logic, as well as adds the necessary
includes to the background jobs to start keeping track of job status.

In a future commit, we will add a new API endpoint that is able to
fetch the status of the plaid jobs given a job id. We will also
need to return job IDs from the existing set_access_token API.