Plaid APIs - SetAccessToken returns Job IDs
This refactors the Plaid#set_access_token endpoint to return the
job_ids of the enqueued jobs that initialize the plaid_item. These
jobs are the get_item_details job and the sync_accounts_job. When
the set_access_token endpoint is called with a new public token,
the plaid item is created and then these two jobs are enqueued
instead of them running async since they make a call to an external
service (Plaid).

The ItemCreate service has been updated to return the job ids as a
hash. The controller will convert the keys in this has to camelCase
to align with API best practices.