Plaid ItemInitializationJobStatuses Service
This adds a new service that is responsible for fetching and
normalizing the status of the two background jobs that enqueue when
a user links a new Plaid Item to a bank.

The queued, running and retrying statuses are normalized to
"pending" while failed is "failed" and complete is "completed".
These are the only statuses that the front end needs to communicate
to the user that work is happening or has stopped.