Add plaid job statuses endpoints
This commit adds a new plaid API endpoint to fetch the status of
the Item initialization jobs. These jobs are the GetItemDetailsJob
and the SyncAccountsJob which enqueue after a successful Plaid
Link.

We also update the tests for the set_access_token endpoint to fix a
typo and update the error message.