Subcategory: Remove name unique constraint
This removes the uniqueness constraint on the subcategory name
field. It's no longer needed as we are now working with subcategory
IDs instead of passing names between the frontend and backend.

We will also be adding new subcategories that have non-unique names
across categories. Categories still, and always will have a
uniqueness constraint on name.