Add Plaid CategoryMapper
This commit adds a new model for mapping Plaid categories to our
applications categories and subcategories.

The model instance will load all categories and subcategories into
memory in an effort to be more efficient in lookups. Transactions
will be added in the magnitude of thousands, so we need to avoid
many database lookups. Since the count of categories and
subcategories will be significantly lower (in the order of less
than 100), then it makes sense to load them all into memory when
importing new transactions.