Add category_mapper to plaid transaction create service
This updates the Plaid::Transaction::Create service to use the new
Plaid::Category mapper to map Plaid transaction categories to our
applications categories.

The create service will create a new instance of the mapper service
if it's not passed in, but if creating many transactions, it's more
efficient to pass in an instance of the category_mapper because it
reduces database calls. The category_mapper pulls all categories
and subcategories from the database into memory rather than make a
database query per transaction created.