Plaid Transaction Sync service: Add category mapper
This adds a category_mapper option to the Plaid Transaction Sync
Service. This category_mapper maps Plaid personal finance
categories to our app's categories. If syncing transactions for
multiple plaid items, it's more efficient to pass in an instance of
the category_mapper than for the sync_service to create it for each
item. This is because the category mapper loads all categories and
subcategories into memory to efficiently do the lookups.