Add Plaid Transaction Modify Service
This commit adds a service responsible for modifying transactions
when Plaid has a modified transaction in their sync. Only the
amount, date, description, pending and category can be updated.

This includes a category_mapper option in the service. The
category_mapper maps Plaid personal finance categories to our
app's categories. If syncing transactions for multiple Plaid
items or accounts, it's more efficient to pass in an instance
of the category_mapper than for the modify service to create
it each time. This is because the category mapper loads all
categories and subcategories into memory to efficiently perform
the lookups.