Plaid Transaction SyncService: Modified transactions
This commit hooks up the Plaid Transaction SyncService to the
ModifiedService by implementing all the logic necessary to modify a
transaction if needed. This typically happens if a transaction was
pending and is finalized, or if an amount changes due to gratuity,
etc...

If the plaid_transaction_id is not found in our database, then we
log it and skip it as to not hold up any other transactions from
syncing.