Add Plaid Item Audits table
This creates a new audit table for the Plaid Item. Since we are
billed per item on the Plaid service, we need a way to keep track
of the Items we create and their products. In a future commit, we
will add hooks to the Plaid Item model to create audits on create,
update and destroy.