Plaid TransactionSync: handle no_account error
This commit implements logic to handle a Plaid::ApiError code of
NO_ACCOUNT. This occurrs when we call transactions_sync on an item
that doesn't have any relevant accounts for transactions. For
example, if a bank is linked and only has investment accounts, then
transactions aren't relevant to this item, and therefore calling
transactions_sync on the item will result in a NO_ACCOUNTS error.

When this happens, we simply log the item_id and return early. In
the future, we'd probably want to figure out a way to not even call
transactions_sync if it's not relevant, but I dont see any
attributes on the item itself that we can reference to determine if
the Item uses transactions or not. So for now, we'll log it as a
warning.

We also update the Plaid API transactions library to just raise the
Plaid::ApiError instead of a custom transactionSync error. This
wasn't adding any value and just made the code more complicated.