Plaid: SyncAccounts Service
This commit refactors the create account and update account logic
out of the SyncAccounts service. The syncAccounts service will be
responsible for fetching the Plaid API data, and then syncing the
accounts whether they exist or not.

In a future commit, we will move the logic from the
get_item_details job into this service since the item data comes
with the account fetch.