Plaid ItemCreate: Refactor to only sync accounts
This commit refactors the Plaid ItemCreate service to only enqueue
the SyncAccounts job which will pull the Item details and sync to
the database. Now that we only enqueue one job, we also refactor
the Plaid Controller set_access_token endpoint to only return a
single job id.

We also rename the item_initialization_job_statuses endpoint to
item_initialization_job_status and update the parameter list to
be a singlular job_id.

This change simplifies a lot of code since we will sync item data
and account data in one job and one Plaid API call.