Plaid Item: Add Sync Service
This commit adds a new service responsible for syncing Plaid Items
locally. The local Plaid item is passed in on initialization, and
then the call method accepts the remote Plaid Item data object.

The institution details require a separate API call. For this
reason, we will only fetch and update the institution details if
the institution_ids don't match between the local data and remote
data.