Plaid::SyncAccounts - Add Item sync
This commit adds logic to the Plaid::SyncAccounts Service to start
syncing Plaid Item details. The Plaid Item details are sent in the
get accounts response, so let leverage this to sync the Item while
we perform the accounts sync.

This saves us an API call to Plaid just to retreive the Item
details. We also can take advantage of knowing that the item
instutution name will be set by the time the accounts are created,
unlike the previous setup where the get item details job was
enqueued at the same time of accounts sync. A future commit will
remove the get item details job since it's now handled by the
AccountsSync service.