Plaid Item Sync Service: remove consented data scopes
This commit removes the consented data scopes field from the Plaid
Item sync service. This field is no longer present when we switched
from Item Get to using the Item object in the get Accounts API
call.

The Item get API returns a Plaid::ItemWithConsentFields object.
The Accounts get API returns a Plaid::Item object. We'll keep the
database column in place for now, but I don't believe we will be
needing this data anyway.