Plaid Account Create Service duplication error
This commit adds logic to the Plaid Accounts CreateService to
handle accidentally attempting to add a duplicate account with the
same plaid_account_id. This could happen for a multitude of reasons
such as concurrency or a race condition. Either way, we don't want
to raise an exception when this happens. We want the service to
gracefully log the error message and return back to the caller.

We also fixed a bug where the balance could be unavailable on the
account. In this case, we want the current_balance attribute to be
set to nil.