Plaid Api: Transaction - Add accounts to response
This adds the Accounts object to the transactions_sync response.
The accounts object was always included in the response from Plaid,
but we haven't needed it until now so it was filtered out.

The accounts object will only be populated from the first page of
transactions results as that should contain all the accounts that
have transactions updates.