Plaid Transactions Sync: Extract and account logic
This commit extracts the Plaid Transactions Sync Service logic into
separate handler classes. The Sync service was getting very big and
each of the added, modified and removed were clearly standalone
methods that could easily be extracted.

We also add logic to the Transaction Sync Added Handler to create
an account if it doesn't exist locally and we have the Plaid
Account data from the transaction sync response.