Add Sidekiq Cron and schedule Plaid TransactionSync
This commit adds the sidekiq-cron gem so that we can schedule
periodic background jobs.

We also take this opportunity to create a separate sidekiq
initializer and move the relevant configurations out of the
application config file.

The first job scheduled is the Plaid::SyncTransactionsJob which
will execute every 5 minutes to ensure transactions are synced on
Plaid Items that need a transaction sync.