Update ruby version and Dockerfile argument
This updates the app ruby version to 3.4.2. All tests pass with
this version.

We also modify the Dockerfile to dynamically pull the ruby-version
from a build time argument. The build_and_push script is updated to
pull this version from the .ruby-version file set by rbenv and pass
it in as a build argument.