Plaid transaction sync to 12 hour period
This updates the Plaid::SyncTransactionsJob to sync transactions
every 12 hours instead of 6 hours. Plaid doesn't update the
transactions more than 2 times per day, so every 12 hours is
sufficient. We also have 200 free transaction API calls to utilize
and we can double the use of them if we double the sync period.