Transaction model: add uncategorized? method
This commit adds a small helper method to the transaction model to
easily check if a transaction is uncategorized.

We also add save to the apply_categorization_rule so that it
persists the category to the transaction. To ensure we would have
caught this bug, .reload is added to the transaction in the
minitest.