Plaid Transaction Amount Mapper
This adds a new mapper class for mapping a Plaid::Transaction
amount to the amount stored in the local transaction. Certain
Plaid accounts will have their transactions come in as negated when
in reality, it should be the other way around. For example,
spending on a credit card should be negative as that is money you
are spending. Credit card payments should register as positive as
the account balance is getting paid off.

We connect this new mapper to the Plaid Transaction CreateService
to start recording amounts correctly. We will need to run a script
to negate the amounts of all of the existing transactions since
they were created before the negation logic.