Account: Remove statement directory parameters
This removes the statement_directory attribute from being used when
creating a manual account. The statement_directory will no longer
be used as users will need to manually upload their statements from
the browser instead of them being synced to the server.

The column will remain in place for now, but a future migration
will eventually remove the column.