Refactor Accounts API
This commit refactors the accounts API mainly to allow updating the
manual_account_provider field. Prior to this change, the business
logic was such that once the account was created, the
account_provider could not be changed. Now that we have Plaid
linked accounts, the business logic needs to change to allow
updating an existing account with a manual_account_provider.

The main reason for this is because Plaid linked accounts will have
the opportunity to have a manual_account_provider added on. When
this happens, users will be able to manually upload statements that
Plaid was unable to fetch. For example, Plaid is only able to pull
the last 3 months of transactions from Capital One. To add
transactions older than 3 months ago, the user can enable
manual_account_provider on the account, and this will allow
statements to be parsed.

We also rename the account_provider attribute to
manual_account_provider to be more specific that it's related to
manual account parsing.

And finally, we add two new attributes to the AccountSerializer:
manualAccountEnabled and plaidAccountEnabled. These are boolean
values to indicate whether an account has a manual parser enabled
or a plaid linked account respectively.