SupportedAccountsEntity: handle nil provider
Updates the supported accounts entity class_from_provider class
method so that an exception is raised if the provider is nil. This
is necessary because the const_defined? method will raise a name
error if the provider portion of the class_name is nil.

A new error type is created to signify this specific error,
EmptyProviderError.

A future commit will update the accounts controller to handle this
new error to let the user know that null manual_account_providers
are not allowed.