Account Serializer: manualAccountProvider attribute
Renames the account serializer attribute manualAccountEnabled to
manualAccountProvider. Since the account provider is now editable
on the account, we should be returning the manual account provider
value when fetching accounts. This way, the user can confirm the
value is correct.

We also update the accounts controller test to ensure this value is
returned correctly.