SupportedAccountsEntity: handle lowercase provider
Updates the logic in the SupportedAccountsEntity
class_from_provider class method to handle when a lowercase
provider is passed in. This was causing a NameError exception to
get raised when the lowercase provider was passed into
Object.const_defined? because Ruby expects all constants to start
with a capital letter.

So, if a user passes in a lowercase provider, we will return an
InvalidParserError.