AccountsController: move error handling to controller
This refactors the Accounts Controller update and create endpoints
so that the error handling is happening in the controller.

First, we remove the InvalidParserError from the
account_services/create service and allow it to bubble up to the
caller. The controller will then convert this into an unprocessable
entity error. The update method also handles this error.

Second, we handle the EmptyProviderError which happens when the
user attempts to create a manual account without a
manual_account_provider value.

Third, we handle the ManualProviderNullificationError in the update
endpoint. This occurrs when the user attempts to nullify the
manual_account_provider on an existing manual account.

We also take this opportunity to split the accounts controller
minitests into different files based on the controller action. Now
the index, create, update and destroy tests are nested under the
accounts_controller test directory.